#1500 - RepositoryProperties not moved for Tests
- fix imports for RepositoryProperties
